Add link popup operability to shopping_cart
The start of the work was present; however, testing
showed that this feature did not work for the
shopping cart page.

Without some of this change, closing out the link
resulted in a blank screen with the ocntent being
hidden through css. Clearing the css left the
page "recentered" so after the colorbox has been
closed, the screen is recentered to place the
original content on screen.