Incorporated/updated/modified auto-installer.
Have added an installer that basically keeps the
admin user aware of what version is installed
by the process of new versions including an updated
file to force the version change. This simplifies
plugin maintenance by not necessarily having to
update a line of code, just create a new file and
use best practices to keep the software up-to-date.
So, this version 2.1.0 is identified for a number
of reasons yet to be listed, but it is an improvement
for general usage over previous versions.