Update installer to not operate on logoff
If the update files were added to the system and
the next click action was the logoff, then the
installer would have executed without visually
notifying the admin.  This lack of notification
is now addressed.

Set the default for version checking to true on
install using the version checking software most
recently applied and such that the checking for
plugin versions is independent of the setting for
ZC overall.
Also converted portions to being PHP 7.2 compatible.
Moved some of the logic for version updating to
an outer check so that resources are not wasted
where the result is not used.