Reworked the options
Changed the options "display" to be less of an
echo condition and more of a present the value to
the browser and where necessary perform php code
execution (echo constants).

This makes it a little easier to see where/when
things like quotes are needed to express a value
and it is also easier for a code editing tool to
see what is going on.