Added same image reload code to script.
Based on discussion, observation, and research,
zen_colorbox while being responsive has required
a little assistance to initially display in a
responsive manner.  With changes made in the
base code, a minor work around was needed and is
documented in the issue log for the software at
issues/158.

With this change, the image will load at 95% the
width of the screen not to be stretched out based
on the image details.  This may result in needing
to scroll down in order to see the remainder of the
image.  This aspect is intentional so that on
mobile devices it is possible to zoom in on the
image rather than have it resize to try to fit the
outer edge of the colorbox within the screen.  This
also causes the image to haven no height which is
problematic.

It is possible to restore resizing on each adjustment
but currently requires a code change in the
jquer.colorbox-min.js file.