Update instructions for ZEN_COLORBOX_PLUGIN_CHECK
The auto installation of the software smartly applies whether Zen Colorbox checks the ZC website for updates or not.  This is done at the point of installation using the define `ZEN_COLORBOX_PLUGIN_CHECK`.  The instructions, however, are viewable from a local computer where PHP is not incorporated.  As a result, the instructions were not showing the information expected.  Therefore, to support someone manually installing the plugin, the instructions are updated to implement the plugin check and to show the quotes around the value.

This fixes #8.