Update 2_1_0.php installer to correct SQL
In review of the installer for other reasons and actually in comparison to another plugin that uses a similar installer, noticed that if the function `zen_page_key_exists` is not present that the SQL statement was not properly formed to produce an appropriate response.  This would only be an issue for a newly installing the plugin or in upgrade from a version prior to this version *AND* if the function did not exist which was added to Zen Cart at or before Zen Cart 1.5.0.